refactor(cli): getWorldFactoryContracts
